+++ components/HardwareIOPlus/test_configHardwareIOPlus.m
@@ -1,0 +1,14 @@
+[cPath, cName, cExt] = fileparts(mfilename('fullpath'));            
+
+% Add mic library
+addpath(genpath(fullfile(cPath, '..', '..')));
+
+cPathConfig = fullfile(...
+    cPath, ...
+    'config-default-stores.json' ...
+);
+        
+config = ConfigHardwareIOPlus(cPathConfig);
+
+
+