+++ devices/Keithley6482/APIVKeithley6482.m
@@ -213,12 +213,7 @@ classdef APIVKeithley6482 < InterfaceKeithley6482
         
         function d = read(this, u8Ch)
             d = this.dMean + this.dSig*randn(1);
-        end
-            
-        
-        
-
-        
+        end  
 
     end