+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.0-alpha.24
+
+### Config
+- Deprecated and it became ConfigHardwareIOPlus for more verbosity
+
+### HardwareIOPlus, Keithley6517A, Keithley6482
+- Updated to use ConfigHardwareIOPlus instead of deprecated Config
+
 # 1.0.0-alpha.23
 
 - Moved tests into component and devide folders so they are easier to find.  Now a component or device folder contains its tests.